backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
kafka-python==2.0.2
psycopg2-binary==2.9.9